+++ databases/alembic/versions/8b36095c6cf9_add_math_type_input_verification.py
@@ -19,8 +19,20 @@ depends_on = None
 def upgrade():
     with op.batch_alter_table("math") as batch_op:
         batch_op.add_column(sa.Column('max_difference', sa.FLOAT))
+        batch_op.add_column(sa.Column('dry_bulb_t_id', sa.FLOAT))
+        batch_op.add_column(sa.Column('dry_bulb_t_measure', sa.FLOAT))
+        batch_op.add_column(sa.Column('wet_bulb_t_id', sa.FLOAT))
+        batch_op.add_column(sa.Column('wet_bulb_t_measure', sa.FLOAT))
+        batch_op.add_column(sa.Column('pressure_pa_id', sa.FLOAT))
+        batch_op.add_column(sa.Column('pressure_pa_measure', sa.FLOAT))
 
 
 def downgrade():
     with op.batch_alter_table("math") as batch_op:
         batch_op.drop_column('max_difference')
+        batch_op.drop_column('dry_bulb_t_id')
+        batch_op.drop_column('dry_bulb_t_measure')
+        batch_op.drop_column('wet_bulb_t_id')
+        batch_op.drop_column('wet_bulb_t_measure')
+        batch_op.drop_column('pressure_pa_id')
+        batch_op.drop_column('pressure_pa_measure')