+++ mycodo/mycodo_flask/page_routes.py
@@ -1011,13 +1011,16 @@ def page_math():
 
     form_add_math = forms_math.MathAdd()
     form_mod_math = forms_math.MathMod()
+    form_mod_multi = forms_math.MathModMultiInput()
+    form_mod_humidity = forms_math.MathModHumidity()
     form_mod_verification = forms_math.MathModVerification()
 
     # convert dict to list of tuples
     choices = []
     for each_key, each_value in choices_input.iteritems():
         choices.append((each_key, each_value))
-    form_mod_math.inputs.choices = choices
+    form_mod_multi.inputs.choices = choices
+    form_mod_verification.inputs.choices = choices
 
     math_templates = []
     math_path = os.path.join(
@@ -1036,10 +1039,12 @@ def page_math():
         elif form_mod_math.mod.data:
             math_type = Math.query.filter(
                 Math.id == form_mod_math.math_id.data).first().math_type
-            if math_type == 'verification':
+            if math_type == 'humidity':
+                utils_math.math_mod(form_mod_math, form_mod_humidity)
+            elif math_type == 'verification':
                 utils_math.math_mod(form_mod_math, form_mod_verification)
             else:
-                utils_math.math_mod(form_mod_math)
+                utils_math.math_mod(form_mod_math, form_mod_multi)
         elif form_mod_math.delete.data:
             utils_math.math_del(form_mod_math)
         elif form_mod_math.order_up.data:
@@ -1059,6 +1064,8 @@ def page_math():
                            display_order=display_order,
                            form_add_math=form_add_math,
                            form_mod_math=form_mod_math,
+                           form_mod_multi=form_mod_multi,
+                           form_mod_humidity=form_mod_humidity,
                            form_mod_verification=form_mod_verification,
                            input=input_dev,
                            math=math,